+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Most recent version is listed first.
 
 
 ## v0.0.1
+- use frankban/quicktest: https://github.com/komuw/shifta/pull/6
 - handle reads efficiently: https://github.com/komuw/shifta/pull/3
 - use uint64 instead of int64: https://github.com/komuw/clog/pull/1
 - add a commitLog datastructure.